fix(gate): apply the slice-vs-whole-target reasoning to ALL THREE ledgers
Round-1 audit, two blockers. I gave GUARD 3's TARGET_FLOORS the "a whole-target
expectation cannot describe a slice of one" treatment and applied it to exactly
one of three ledgers. GUARD 7's NOLAUNCH_ACK and GUARD 2's EXPECTED_SKIPS are
the same kind of expectation and got none of it.

REPRODUCED on this branch, both with every selected test PASSING:
  --files scripts/tests/test_ship_detect_role.py
    -> 12 passed, `GUARD 7: scripts/tests intercepted NOTHING`, exit 1
       (green iff the selection happened to include one of THREE launcher-seam
        files out of 181 — a coin flip, on 63% of the suite)
  --files scripts/signal/tests/test_search.py
    -> 15 passed, `0 skipped, but 2 of 3 pinned entries apply here`, exit 1
       (green iff it happened to include test_pg_type_compat.py, of 920 tests)
Both printed remediation advice that would have disabled a guard repo-wide or
reddened the FULL gate if followed.

ONLY the whole-target half is suspended, and suspension is ANNOUNCED:
  GUARD 7  REQUIRED "this target intercepts >=1"  -> SUSPENDED
           PERMITTED "no OTHER target may"        -> KEPT (per-observation)
  GUARD 2  TOTAL "skips == applicable pins"       -> SUSPENDED
           UNPINNED "no skip unexplained"         -> KEPT (per-observation)
Plugin-marker counts are kept too — they ask whether the guard loaded for THIS
invocation, which a slice can answer. A new SUMMARY block lists every suspended
expectation; a guard that quietly stops applying is the #276 shape.

🔴 AND THE SUITE WAS STRUCTURALLY BLIND TO BOTH: every scoped fixture built its
runner with ack=[] and an empty EXPECTED_SKIPS, so both ledgers were inert
across all 45 tests. runner_patch gains `expected_skips`, and six new tests
build the combined state — each paired with a non-scoped control on the SAME
copy, because "the scoped run is green" is also satisfied by deleting the guard.

Also fixed:
- three zero-test invocations (--check-targets/--check-floors/--check-suites)
  printed `SCOPE: FULL` + `RESULT: PASS (exit=0)` in ~2s having collected
  nothing. New `NONE` state; NONE wins over narrowing, because a scope state is
  a claim about coverage.
- --files accepted a SHELL_TEST (scripts/tests/*.sh sits under the scripts/tests
  DIRECTORY target), handed it to pytest, collected 0, and called it "a
  collection error or an import breakage". Refused by family, naming it.
- gate.sh refused only DEVRC_TARGETS while claiming it was "the ONLY way" the
  tier gets narrowed. DEVRC_GATE_PYTEST_RUNNER/NODE_RUNNER replace the runner
  wholesale — a green verdict off nothing run — and MIN_TESTS overrides the
  floor. All four refused; DEVRC_GATE_ALLOW_AMBIENT is the test-only seam.
- scoped-tests.sh refuses an ambient DEVRC_TARGETS: it shrinks the mapper's
  universe, so the exit-4 refusal printed a wrong DIAGNOSIS.
- CLAUDE.md said "every GUARD still applies" — false three ways.
- scope_line() took the FIRST match while gate.sh takes the LAST.

🔴 SPLIT INTO FIVE FILES. --dist loadfile pins a file to ONE worker, so the
53-test file was a serial critical path inside the biggest target, on the
bottleneck tier — self-refuting in this PR. MEASURED: 187.6s wall (one file,
load 58) -> 59.5s (five files, -n 4 --dist loadfile, load 79.75), ~3.1x, at
higher load. Shared fixtures: scripts/testlib/scoped_harness.py.
